Move ring/openssl to a 'testing' package
