+++ aws-lc-rs-testing/benches/cipher_benchmark.rs
