+++ aws-lc-rs-testing/benches/data/cipher_aes_128_ctr.txt
