+++ aws-lc-rs-testing/benches/digest_benchmark.rs
