+++ aws-lc-rs-testing/benches/ed25519_benchmark.rs
