+++ aws-lc-rs-testing/benches/hmac_benchmark.rs
